Handle transport closed in SSL
For some reason I thought I would get SSL_ERROR_ZERO_RETURN if my
transport was closed in the middle of an SSL_* call, but how would
OpenSSL even know if there was no close_notify being received? It
absolutely needs to be handled in the underlying recv/send syscalls.