Add error_code to filewatcher callback
If I regenerate the whole output directory for my website, it gets
deleted, so all directory watches receive IN_IGNORED. But we try to
rewatch immediately, which fails (because the directory is not there
yet) and we never really recover from that. Now I bubble up that error
and remove the watch and the file from the cache completely, so on the
next request all of it is reinitialized.

Another solution would be to go up the directory tree and try to watch
parent directories until something succeeds and then rebuild the
watch-tree automatically when the corresponding child directories are
recreated. I think this is tricky to build and likely buggy for a while
and not much better than what I have now. Additionally there are still
ways this could file (e.g. when the number of inotify fds is exceeded)
and it would still be useful to be able to bubble up errors. So maybe I
will do this some day.