Fix pending bytes not being sent when SSL op has completed
This is a case of a faulty behaviour that was never actually a problem
(all major browser, curl, wrk, hey, ab, etc. all worked fine), but it is
wrong and to fix it I had to take the whole thing apart.
I do not like the current version *at all*. It's an implicit state
machine again, but enumerating the states and making it explicit just
makes it worse, I think (because many are the same). I don't feel good
committing this, but it's better to be correct than to be pretty.
Maybe I'll come up with a nice way some other time. This bug was
haunting me for many months and at least it's off the table now.

Also I noticed some other potential bugs (i.e. we would do a send even
though there was an SSL_ERROR_SSL for example), which I have fixed as
well.