Add a bridge to make HTTP requests from a different thread
I don't know how much I like this idea right now, but it's enough for
what I want to do and it's not too horrible, I think. It couples with
nothing (stricly on top of existing code) and has the interface I want,
so I might keep it even if it is not pretty.