Add ACME client (and automatic HTTPS)
This took a LONG time and it's a huge commit, because I cannot break it
up further. The previous few commits were all essentially also part of
this.
I started this thinking that it would be really cool if it didn't take
that much time and I made a simple stand-alone ACME client work in a few
days. Integrating it into htcpp took so long that I had to take a few
breaks inbetween. I am so happy this works and is done now.