Add max concurrent connections limit
I put this off for so long, because I couldn't come up with an
implementation that doesn't look weird. I chose to just go with the
weird way, because I'd rather have the feature and not.
The weird part is that I have to pass a reference to the connection
counter into the Session, because it needs to live in the Server object,
which does not keep track of the Sessions.

Also this is tested only so-so. It's surprisingly hard to convince a
traffic generator to respect the concurrency you specify. hey just opens
a bunch more connections sometimes (looks pretty random too). wrk clamps
down to weird numbers (e.g. 180, then 192). It works .. pretty okay at
least. I can't say if a limit of 200 actually limits to 200 connections
and not 190 or 210 (of course I can tell from the code, but not from
testing), but it will not deny new connections at 150. Some debug
printing of the connection counter and correlating with /proc/(pidof
trafficgenerator)/fd looks fine too. And it does reset to zero when it
should.