Try out not using the slot map
Normally I save an integer in the SQE user data that I use as a key in a
"slot map" data structure to map to completion handler callbacks.
I considered in the past to simply store a pointer to my data structure
in that user data. I didn't like it, because I would have to use a raw
new/delete and I would store a pointer potentially in some kernel
memory. Maybe I'll somehow lose that pointer and never free my handlers?
It felt dangerous, but looking at it again I keep thinking that the slot
map is needlessly complicated, so I tried it out and the difference is
neglibile. I tested many times, but the performance is pretty much
exactly the same (57k req/s for both versions). So I'd rather keep the
safer approach.